Rename MarkerUpdate > Location, MarkerManager > UserDataManager.
